+++ terraform/modules/aws-load-balancer-controller/outputs.tf
@@ -1,0 +1,19 @@
+output "iam_role_arn" {
+  description = "ARN of the IAM role for the AWS Load Balancer Controller"
+  value       = aws_iam_role.alb_controller.arn
+}
+
+output "iam_policy_arn" {
+  description = "ARN of the IAM policy for the AWS Load Balancer Controller"
+  value       = aws_iam_policy.alb_controller.arn
+}
+
+output "helm_release_name" {
+  description = "Name of the Helm release"
+  value       = helm_release.aws_load_balancer_controller.name
+}
+
+output "helm_release_version" {
+  description = "Version of the Helm chart deployed"
+  value       = helm_release.aws_load_balancer_controller.version
+}